pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logistics-data-platform"
version = "0.1.0"
description = "Production-grade logistics and delivery analytics platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Data Engineering Team" }
]
keywords = ["logistics", "data-engineering", "bigquery", "airflow", "dbt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "apache-airflow>=2.8.0",
    "apache-airflow-providers-google>=10.0.0",
    "dbt-bigquery>=1.7.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "google-cloud-bigquery>=3.14.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-pubsub>=2.19.0",
    "great-expectations>=0.18.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "faker>=22.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.28.0",
    "jupyterlab>=4.0.0",
]

[project.scripts]
logistics = "data_generation.main:app"

[tool.hatch.build.targets.wheel]
packages = ["data_generation", "api", "feature_store"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.isort]
known-first-party = ["data_generation", "api", "feature_store"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=data_generation --cov=api --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["data_generation", "api", "feature_store"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
